+++ src/main/java/com/example/let_server/domain/menuAllergy/repository/MenuAllergyRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 
 public interface MenuAllergyRepository {
     MenuAllergy save(MenuAllergy menuAllergy);
-    List<Allergy> findAllergyByMealId(Long mealId);
+    List<Allergy> findAllergyByMealId(Integer mealId);
     void saveAllBatch(Long menuId, List<Long> allergyIds);
 }
